Refactor auth configuration and separate server-side user retrieval
