+++ src/lib/auth/actions.ts
@@ -5,34 +5,19 @@ import { signIn } from "next-auth/react"
 import { toast } from "sonner"
 
 import { auth } from "@/lib/auth/auth"
-import { User, type IUser } from "@/lib/models"
-import { connectDB } from "@/lib/db"
 
 /**
  * Get the current user
  * @returns The current user
  */
-export async function getCurrentUser(): Promise<Omit<IUser, "password"> | null> {
+export async function getCurrentUser() {
   try {
-    await connectDB()
     const session = await auth()
     if (!session?.user) {
       return null
     }
 
-    const user = await User.findOne({
-      _id: session.user.id,
-      archivedAt: null,
-    })
-
-    if (!user) {
-      return null
-    }
-
-    // exclude password
-    const { password: _password, ...userWithoutPassword } = user.toObject()
-
-    return userWithoutPassword
+    return session.user
   } catch (error) {
     console.error("Error getting current user:", error)
     return null
